upi/requirements.txt
flask
numpy
pandas
scikit-learn
shap
networkx
celery
redis
pyotp
cryptography
flask-login
python-fido2

